+++ tests/integration/cloudera/test_serializing_avromodel.py
@@ -3,9 +3,10 @@ from pathlib import Path
 import pytest
 from pydantic import Field, BaseModel
 
+from wunderkafka.serdes.headers import ConfluentClouderaHeadersHandler
 from wunderkafka.time import now
 from wunderkafka.tests import TestProducer, TestHTTPClient
-from wunderkafka.serdes.avro import AvroModelSerializer, ConfluentClouderaHeadersHandler
+from wunderkafka.serdes.avro import AvroModelSerializer
 from wunderkafka.serdes.store import AvroModelRepo
 from wunderkafka.schema_registry import SimpleCache, ClouderaSRClient
 from wunderkafka.producers.constructor import HighLevelSerializingProducer
@@ -37,7 +38,7 @@ def ts() -> int:
 
 @pytest.fixture
 def value_answer() -> bytes:
-    return b'\x01\x00\x00\x00\x00\x00\x00\x06\x9c\x00\x00\x00\x01\xcc\xb8\xeb\xa6\x80_\x08test\x0cstring\x0cstring\x04NA'
+    return b'\x01\x00\x00\x00\x00\x00\x00\x06\x9c\x00\x00\x00\x01\xcc\xb8\xeb\xa6\x80_\x08test\x0cstring\x0cstring\x04NA'  # noqa: E501
 
 
 @pytest.fixture