+++ tests/smoke/test_factories.py
@@ -1,8 +1,4 @@
-import sys
-
-import pytest
-
-from wunderkafka.factories import AvroConsumer, AvroProducer, ConsumerConfig, ProducerConfig, AvroModelProducer
+from wunderkafka.factories.avro import AvroConsumer, AvroProducer, ConsumerConfig, ProducerConfig, AvroModelProducer
 from wunderkafka.config.schema_registry import SRConfig