+++ wunderkafka/compat.py
@@ -1,0 +1,6 @@
+import sys
+
+if sys.version_info >= (3, 10):
+    from typing import ParamSpec
+else:
+    from typing_extensions import ParamSpec