+++ wunderkafka/consumers/constructor.py
@@ -8,10 +8,12 @@ All moving parts should be interchangeable in terms of schema, header and serial
 """
 
 import datetime
-from typing import Any, List, Union, Optional
+from typing import Any, List, Union, Optional, TypeVar
 
 from confluent_kafka import Message, TopicPartition
+from confluent_kafka.serialization import SerializationError
 
+from wunderkafka.consumers.types import StreamResult, PayloadError
 from wunderkafka.types import HeaderParser
 from wunderkafka.logger import logger
 from wunderkafka.serdes.abc import AbstractDeserializer
@@ -19,6 +21,23 @@ from wunderkafka.structures import SchemaMeta, SchemaDescription
 from wunderkafka.consumers.abc import AbstractConsumer, AbstractDeserializingConsumer
 from wunderkafka.schema_registry.abc import AbstractSchemaRegistry
 from wunderkafka.consumers.subscription import TopicSubscription
+
+T = TypeVar("T")
+
+
+def choose_one_of(
+    common_deserializer: Optional[AbstractDeserializer],
+    specific_deserializer: Optional[AbstractDeserializer],
+    what: str,
+) -> AbstractDeserializer:
+    chosen_deserializer = specific_deserializer if specific_deserializer else common_deserializer
+    if chosen_deserializer is None:
+        msg = [
+            '{0} deserializer is not specified,'.format(what.capitalize()),
+            'it should be passed via {0}_deserializer or deserializer at least.'.format(what),
+        ]
+        raise ValueError(' '.join(msg))
+    return chosen_deserializer
 
 
 class HighLevelDeserializingConsumer(AbstractDeserializingConsumer):
@@ -29,7 +48,11 @@ class HighLevelDeserializingConsumer(AbstractDeserializingConsumer):
         consumer: AbstractConsumer,
         headers_handler: HeaderParser,
         schema_registry: AbstractSchemaRegistry,
-        deserializer: AbstractDeserializer,
+        deserializer: Optional[AbstractDeserializer] = None,
+        value_deserializer: Optional[AbstractDeserializer] = None,
+        key_deserializer: Optional[AbstractDeserializer] = None,
+        *,
+        stream_result: bool = False,
     ):
         """
         Init consumer with specific dependencies.
@@ -37,12 +60,20 @@ class HighLevelDeserializingConsumer(AbstractDeserializingConsumer):
         :param consumer:            Consumer implementation to receive messages.
         :param headers_handler:     Callable to parse binary headers.
         :param schema_registry:     Schema registry client.
-        :param deserializer:        Message deserializer.
+        :param deserializer:        Common message deserializer for value and key, if set.
+        :param value_deserializer:  Message deserializer for value, if set.
+        :param key_deserializer:    Message deserializer for the key, if set.
+        :param stream_result:       If True, return a complex StreamResult object instead of just a model.
         """
         self.consumer = consumer
         self._header_parser = headers_handler
         self._registry = schema_registry
         self._deserializer = deserializer
+
+        self._value_deserializer = choose_one_of(deserializer, value_deserializer, 'value')
+        self._key_deserializer = choose_one_of(deserializer, key_deserializer, 'key')
+
+        self._stream_result = stream_result
 
     def subscribe(  # noqa: D102,WPS211 # docstring inherited from superclass.
         self,
@@ -78,7 +109,7 @@ class HighLevelDeserializingConsumer(AbstractDeserializingConsumer):
         ignore_keys: bool = False,
         raise_on_error: bool = True,
         raise_on_lost: bool = True,
-    ) -> List[Message]:
+    ) -> List[T]:
         """
         Consume as many messages as we can for a given timeout and decode them.
 
@@ -90,20 +121,19 @@ class HighLevelDeserializingConsumer(AbstractDeserializingConsumer):
         :param raise_on_lost:   If True, check on own clocks if max.poll.interval.ms is exceeded. If so, raises
                                 ConsumerException to be handled in client code.
 
-        :raises KafkaError:     See https://docs.confluent.io/platform/current/clients/confluent-kafka-python/html/index.html#confluent_kafka.KafkaError
+        :raises KafkaError:     See https://docs.confluent.io/platform/current/clients/confluent-kafka-python/html/index.html#confluent_kafka.KafkaError  # noqa: E501
 
         :return:                A list of Message objects with decoded value() and key() (possibly empty on timeout).
         """  # noqa: E501
         msgs = self.consumer.batch_poll(timeout, num_messages, raise_on_lost=raise_on_lost)
-        return self._decoded(msgs, ignore_keys=ignore_keys, raise_on_error=raise_on_error)
+        return self._decoded(
+            msgs,
+            ignore_keys=ignore_keys,
+            raise_on_error=raise_on_error,
+        )
 
-    def _decoded(
-        self,
-        msgs: List[Message],
-        *,
-        ignore_keys: bool,
-        raise_on_error: bool,
-    ) -> Any:
+    def _decoded(self, msgs: List[Message], *, ignore_keys: bool, raise_on_error: bool) -> List[T]:
+        results: List[StreamResult] = []
         for msg in msgs:
             kafka_error = msg.error()
             # Not sure if there is any need for an option to exclude errored message from the consumed ones,
@@ -114,23 +144,58 @@ class HighLevelDeserializingConsumer(AbstractDeserializingConsumer):
             if kafka_error is not None:
                 logger.error(kafka_error)
                 if raise_on_error:
+                    # Even PyCharm stubs show that it is inherited from an object, in fact it is a valid Exception
                     raise kafka_error
 
             topic = msg.topic()
-            msg.set_value(self._decode(topic, msg.value()))
+
+            raw_key_value = msg.key()
+            decode_key_ok = True
             if ignore_keys:
                 # Yes, we lose information, but it is necessary to not get raw bytes
                 # if `.key()` will be called in client code later.
                 msg.set_key(None)
             else:
-                msg.set_key(self._decode(topic, msg.key(), is_key=True))
-        return msgs
+                try:
+                    decoded_key = self._decode(topic, msg.key(), is_key=True)
+                # KeyDeserializationError is inherited from SerializationError
+                except SerializationError:
+                    decode_key_ok = False
+                    logger.error("Unable to decode key from bytes: {0}".format(raw_key_value))
+                    if not ignore_keys:
+                        raise
+                else:
+                    msg.set_key(decoded_key)
+
+            try:
+                decoded_value = self._decode(topic, msg.value())
+            except (SerializationError, ValueError) as exc:
+                logger.error("Unable to decode value from bytes: {0}".format(msg.value()))
+                if not self._stream_result:
+                    raise
+                value_error = str(exc)
+                if decode_key_ok:
+                    error = PayloadError(description=value_error)
+                else:
+                    message = "Unable to decode key (topic: {0}, key payload: {1})".format(topic, raw_key_value)
+                    error = PayloadError(description=message)
+                results.append(StreamResult(payload=None, error=error, msg=msg))
+            else:
+                msg.set_value(decoded_value)
+                if self._stream_result:
+                    results.append(StreamResult(payload=decoded_value, error=None, msg=msg))
+
+        to_return = results if self._stream_result else msgs
+        return to_return
 
     # Todo (tribunsky.kir): arguable: make different composition (headers, SR & deserializer united cache)
     def _decode(self, topic: str, blob: Optional[bytes], *, is_key: bool = False) -> Any:
         if blob is None:
             return None
 
+        deserializer = self._get_deserializer(is_key)
+        if deserializer.schemaless:
+            return deserializer.deserialize('', blob)
         # Header is separate in the sake of customization, e.g., we don't have SR and put schema directly in a message
         parsed_header = self._header_parser(blob)
         schema_meta = SchemaMeta(
@@ -144,4 +209,7 @@ class HighLevelDeserializingConsumer(AbstractDeserializingConsumer):
         # - copy the whole tail
         # - have implicit offset as if we read buffer when extracting header
         # so dealing with implicit offset and the whole binary string
-        return self._deserializer.deserialize(schema.text, blob, seek_pos=parsed_header.size)
+        return deserializer.deserialize(schema.text, blob, seek_pos=parsed_header.size)
+
+    def _get_deserializer(self, is_key: bool) -> AbstractDeserializer:
+        return self._key_deserializer if is_key else self._value_deserializer