+++ wunderkafka/schema_registry/clients/subject.py
@@ -1,0 +1,1 @@
+# ToDo (tribunsky.kir): re-implement or write adapter for confluent-kafka'