+++ wunderkafka/serdes/string/__init__.py
@@ -1,0 +1,1 @@
+# FixMe (tribunsky.kir): I will burn in hell for that, but for now it's much easier to just wrap everything